refactor: 오브제 수정 퍼사드 패턴 적용
